refactor: migrate build to platformio/espressif32@7.0.1 (IDF 6.x / GCC 15.2)

+++ src/util/ScreenshotUtil.cpp
@@ -14,7 +14,7 @@
 #include "activities/Activity.h"
 
 void ScreenshotUtil::buildFilename(const ScreenshotInfo& info, char* buf, size_t bufSize) {
-  const uint32_t ts = uptime_ms();
+  const unsigned int ts = static_cast<unsigned int>(uptime_ms());
 
   if (info.readerType == ScreenshotInfo::ReaderType::None || info.title[0] == '\0') {
     snprintf(buf, bufSize, "/screenshots/screenshot-%u.bmp", ts);
@@ -147,7 +147,7 @@ bool ScreenshotUtil::saveFramebufferAsBmp(const char* filename, const uint8_t* f
   // Max row size for 528px height (X3) after rotation = 68 bytes; use fixed buffer to avoid VLA
   constexpr size_t kMaxRowSize = 68;
   if (rowSizePadded > kMaxRowSize) {
-    LOG_ERR("SCR", "Row size %u exceeds buffer capacity", rowSizePadded);
+    LOG_ERR("SCR", "Row size %u exceeds buffer capacity", static_cast<unsigned int>(rowSizePadded));
     // Explicitly close() file before calling Storage.remove()
     file.close();
     Storage.remove(filename);
